Refactor for PR
